Fixed search glitch // added page state
Race condition issue present in handleSearch function. setting the state of the zipcode and/or distance did not happen before the function called and ran the getPetsFromApi function. This made it so the search button would need to be clicked twice to actually search the criteria.

Deleted function call from handleSearch, and added zipcode and distance state to useEffect. Glitch resolved.

Added page state, and pg: page to api call. This is the beginning of accounting for additional search results.