Added logic to request addl pages
petfinder api returns 'pages' of pets. Once user has seen all returned pets, they will want to see more.

Completed setup of page state. created logic when setting pets state to determine if query needs to be concat'ed to prevpets.

Created button that triggers this to happen, and at the same time, the setPage is incremented by 1.

If the user changes the search criteria, setPage will set back to 1.

Next step is to delete the button, and make the loadNextPage function run when user reaches bottom of page.