Added proper logic for zipcode and distance
If a zipcode is entered, the default distance is 50 miles. Edits made to my server side routing required I set the distance state to 50.

This change means that all my (distance ?) logic is no longer useful. Since the distance is only used if a zipcode is entered, I was able to change everything to (zipcode ?), including pushing push the zipcode state from App.js to PetList for the filtering done in that component.