Fixed Duplicate results issue/ and duplicate api calls
Tracked down that duplicate results is due to petfinders api service, so I'd need to handle duplicates on my end.

Created filter on petList component to remove duplicates.

Also, corrected VisibilitySensor to on trigger when the element comes in to view. Previously, it would trigger when it came in to view, which would load more pets, pushing the element out of view, which would trigger again.